[Client, Server] navBar 프사 css 수정 / server/image.js 수정

+++ server/controllers/user/image.js
@@ -25,6 +25,31 @@ module.exports = {
         const userInfo = await user.findOne({
           where: { id: refreshVerify.id },
         });
+        if (userInfo.dataValues.userPic) {
+          const decode = decodeURI(userInfo.dataValues.userPic);
+          let path = decode.split('/')[4];
+          s3.deleteObject(
+            {
+              Bucket: 'jurimma.com',
+              Key: path,
+            },
+            (err, data) => {
+              if (err) {
+                throw err;
+              }
+              console.log('deleted', data);
+            }
+          );
+        }
+        userInfo.userPic = req.file.location;
+        await userInfo.save();
+        res.status(201).json({ accessToken, userPic: req.file.location });
+      }
+    }
+    // accessToken 유효
+    else {
+      const userInfo = await user.findOne({ where: { id: accessVerify.id } });
+      if (userInfo.dataValues.userPic) {
         const decode = decodeURI(userInfo.dataValues.userPic);
         let path = decode.split('/')[4];
         s3.deleteObject(
@@ -39,30 +64,7 @@ module.exports = {
             console.log('deleted', data);
           }
         );
-
-        userInfo.userPic = req.file.location;
-        await userInfo.save();
-        res.status(201).json({ accessToken, userPic: req.file.location });
       }
-    }
-    // accessToken 유효
-    else {
-      const userInfo = await user.findOne({ where: { id: accessVerify.id } });
-      const decode = decodeURI(userInfo.dataValues.userPic);
-      let path = decode.split('/')[4];
-      // 기존에 저장되어있던 데이터 s3에서 삭제
-      s3.deleteObject(
-        {
-          Bucket: 'jurimma.com',
-          Key: path,
-        },
-        (err, data) => {
-          if (err) {
-            throw err;
-          }
-          console.log('deleted', data);
-        }
-      );
       userInfo.userPic = req.file.location;
       await userInfo.save();
       res.status(200).json({ userPic: req.file.location });
